Removed functions from the unique interface
Removed functions from the unique interface as this produced an error
with gfortran, as the interface included both functions and
subroutines.  This did not produce errors in ifort.  I am going to
check if this is becuase if this is a feature of the newer standards or
if it is just an ifort extension.